move sub classes to csvcomponents py file
